backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cassandra-twitter-mcp"
version = "0.2.0"
description = "Consolidated Twitter MCP server — Grok synthesis + X API v2 + personal account via twitter-cli"
requires-python = ">=3.12"
dependencies = [
    "cassandra-mcp-auth @ git+https://github.com/DigiBugCat/cassandra-auth.git#subdirectory=python",
    "fastmcp>=2.12.0",
    "httpx>=0.28.0",
    "twitter-cli>=0.8.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
cassandra-twitter-mcp = "cassandra_twitter_mcp.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
